Partial deploy: all, client, server

+++ bash/brahma-deploy.bash
@@ -9,6 +9,15 @@
 #   wait $pid
 # done
 
+projectsServer="brahma-api brahma-chat brahma-notifications"
+projectsClient="hearthy-client-public hearthy-professional-public brahma-admin-public"
+
+projects="all"
+
+if [ -n "$1" ]; then
+  projects=$1
+fi
+
 color_1='\033[0;35m'
 color_2='\033[1;35m'
 color_3='\033[0;90m'
@@ -20,20 +29,26 @@ echo -e "${color_1}${color_2} ____  _____ _____ __    _____ __ __    _____ _____
 |____/|_____|__|  |_____|_____| |_|    |_____| |_| |__|__|__|__| |_| |_____|$color_no"
 
 ./grunt
-for project in "brahma-api" "brahma-chat" "brahma-notifications"; do
-  cd "$project"
-  echo -e "${color_1}Deploying ${color_2}$project$color_no"
-  git push --force DEV develop:master
-  cd ..
-done
-for project in "hearthy-client-public" "hearthy-professional-public" "brahma-admin-public"; do
-  cd "$project"
-  echo -e "${color_1}Deploying ${color_2}$project$color_no"
-  python -c'import time; print repr(time.time())' > .force_push
-  git add .
-  git commit --amend -a -m "deployment" > /dev/null
-  git push --force DEV
-  cd ..
-done
+
+if [ "$projects" == "all" -o "$projects" == "server" ]; then
+  for project in $projectsServer; do
+    cd "$project"
+    echo -e "${color_1}Deploying ${color_2}$project$color_no"
+    git push --force DEV develop:master
+    cd ..
+  done
+fi
+
+if [ "$projects" == "all" -o "$projects" == "client" ]; then
+  for project in $projectsClient; do
+    cd "$project"
+    echo -e "${color_1}Deploying ${color_2}$project$color_no"
+    python -c'import time; print repr(time.time())' > .force_push
+    git add .
+    git commit --amend -a -m "deployment" > /dev/null
+    git push --force DEV
+    cd ..
+  done
+fi
 
 echo -e "${color_1}${color_2}END OF DEPLOYMENT$color_no"
